testing_lists_for_moulinette

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: vrayinch <vrayinch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/28 10:36:43 by vrayinch          #+#    #+#             */
-/*   Updated: 2020/01/18 16:27:23 by vrayinch         ###   ########.fr       */
+/*   Updated: 2020/01/27 16:44:10 by vrayinch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long long	ft_strlen(char const *str)
+size_t	ft_strlen(char const *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len])
